+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added support for SCRAM-SHA-256 and SCRAM-SHA-512 authentication
 * Added Python script tool to calculate latencies and quantiles from canary log for producer, consumer and connection buckets
+* #133: turn off producer retry and add a consumer error metric
 
 ## 0.1.0